+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/chat/modify/PinChatMessage.kt
@@ -1,7 +1,6 @@
 package com.github.insanusmokrassar.TelegramBotAPI.requests.chat.modify
 
-import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types.ChatRequest
-import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types.DisableNotification
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.SimpleRequest
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import kotlinx.serialization.*
@@ -12,10 +11,10 @@ data class PinChatMessage (
     @SerialName(chatIdField)
     override val chatId: ChatIdentifier,
     @SerialName(messageIdField)
-    val messageId: MessageIdentifier,
+    override val messageId: MessageIdentifier,
     @SerialName(disableNotificationField)
     override val disableNotification: Boolean = false
-): ChatRequest, SimpleRequest<Boolean>, DisableNotification {
+): ChatRequest, SimpleRequest<Boolean>, MessageAction, DisableNotification {
     override fun method(): String = "pinChatMessage"
     override fun resultSerializer(): KSerializer<Boolean> = BooleanSerializer
 }