+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/edit/abstracts/EditChatMessage.kt
@@ -1,11 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.requests.edit.abstracts
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types.MessageAction
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.SimpleRequest
-import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
-import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.RawMessage
 
-interface EditChatMessage : SimpleRequest<RawMessage> {
-    val chatId: ChatIdentifier
-    val messageId: MessageIdentifier
-}
+interface EditChatMessage : SimpleRequest<RawMessage>, MessageAction