+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/Common.kt
@@ -6,6 +6,7 @@ typealias InlineQueryIdentifier = String
 typealias UpdateIdentifier = Long
 typealias MediaGroupIdentifier = String
 typealias ForwardSignature = String
+typealias ForwardSenderName = String
 typealias AuthorSignature = ForwardSignature
 typealias CallbackQueryIdentifier = String
 typealias PaymentQueryIdentifier = String
@@ -15,6 +16,7 @@ typealias InvoicePayload = String
 typealias ShippingOptionIdentifier = String
 typealias StartParameter = String
 typealias InlineMessageIdentifier = String
+typealias PollIdentifier = String
 
 val callbackQueryAnswerLength = 0 until 200
 val captionLength = 0 until 1024
@@ -28,6 +30,10 @@ val allowedConnectionsLength = 1 .. 100
 val invoiceTitleLimit = 1 until 32
 val invoiceDescriptionLimit = 1 until 256
 val invoicePayloadBytesLimit = 1 until 128
+
+val pollOptionTextLength = 1 .. 100
+val pollQuestionTextLength = 1 until 256
+val pollOptionsLimit = 2 .. 10
 
 val livePeriodLimit = 60 .. 86400
 
@@ -75,6 +81,8 @@ const val hasCustomCertificateField = "has_custom_certificate"
 const val pendingUpdateCountField = "pending_update_count"
 const val lastErrorDateField = "last_error_date"
 const val lastErrorMessageField = "last_error_message"
+const val votesCountField = "voter_count"
+const val isClosedField = "is_closed"
 
 
 const val photoUrlField = "photo_url"
@@ -124,6 +132,7 @@ const val inputMessageContentField = "input_message_content"
 const val hideUrlField = "hide_url"
 
 
+const val isMemberField = "is_member"
 const val canSendMessagesField = "can_send_messages"
 const val canSendMediaMessagesField = "can_send_media_messages"
 const val canSendOtherMessagesField = "can_send_other_messages"
@@ -180,6 +189,8 @@ const val payloadField = "payload"
 const val vcardField = "vcard"
 const val resultsField = "results"
 const val certificateField = "certificate"
+const val questionField = "question"
+const val optionsField = "options"
 
 const val pointField = "point"
 const val xShiftField = "x_shift"