+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/chat/GroupChat.kt
@@ -1,15 +1,9 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.chat
 
-import com.github.insanusmokrassar.TelegramBotAPI.types.ChatId
-import com.github.insanusmokrassar.TelegramBotAPI.types.ChatPhoto
+import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.GroupChat
 
-interface GroupChat : PublicChat {
-    val allMembersAreAdmins: Boolean
-}
-
-data class GroupChatImpl(
-    override val id: ChatId,
-    override val title: String? = null,
-    override val allMembersAreAdmins: Boolean,
-    override val inviteLink: String? = null,
-    override val chatPhoto: ChatPhoto? = null) : GroupChat
+@Deprecated(
+    "Replaced into another package",
+    ReplaceWith("GroupChat", "com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.GroupChat")
+)
+typealias GroupChat = GroupChat