+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/chat/RawChat.kt
@@ -1,6 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.chat
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
+import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.RawMessage
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable