+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/chat/abstracts/PublicChat.kt
@@ -1,0 +1,9 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts
+
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.RawMessage
+
+interface PublicChat : Chat {
+    val title: String?
+    val inviteLink: String?
+    val pinnedMessage: RawMessage?
+}