+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/abstracts/Message.kt
@@ -1,7 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
-import com.github.insanusmokrassar.TelegramBotAPI.types.chat.Chat
+import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
 import org.joda.time.DateTime
 
 interface Message {