+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/content/PollContent.kt
@@ -1,0 +1,28 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.message.content
+
+import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.Request
+import com.github.insanusmokrassar.TelegramBotAPI.requests.send.SendPoll
+import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
+import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.KeyboardMarkup
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.RawMessage
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MessageContent
+import com.github.insanusmokrassar.TelegramBotAPI.types.polls.Poll
+
+data class PollContent(
+    val poll: Poll
+) : MessageContent {
+    override fun createResend(
+        chatId: ChatIdentifier,
+        disableNotification: Boolean,
+        replyToMessageId: MessageIdentifier?,
+        replyMarkup: KeyboardMarkup?
+    ): Request<RawMessage> = SendPoll(
+        chatId,
+        poll.question,
+        poll.options.map { it.text },
+        disableNotification,
+        replyToMessageId,
+        replyMarkup
+    )
+}